Time: 179 ms (14.51%), Space: 53.8 MB (8.13%) - LeetHub

+++ 0018-4sum/0018-4sum.cpp
@@ -8,7 +8,7 @@ public:
             for(int j = i + 3; j < nums.size(); j++) {
                 int left = i + 1;
                 int right = j - 1;
-                long long newTarget = target - nums[i] - nums[j];
+                long long newTarget = (long long) target - nums[i] - nums[j];
 
                 while(left < right) {
                     if(nums[left] + nums[right] > newTarget) right--;
